pyproject.toml
[project]
name = "launch-cli"
version = "0.3.0"
description = "CLI tooling for common Launch functions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache 2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["click>=8.1.7", "pygithub>=2.1.1", "click-spinner>=0.1.10", "gitpython>=3.1", "semver>=3.0", "Jinja2>=3.1.3", "ruamel.yaml>=0.17.32"]

[project.urls]
Homepage = "https://github.com/nexient-llc/launch-cli"
Issues = "https://github.com/nexient-llc/launch-cli/issues"

[tool.setuptools]
package-dir = {launch="src/launch"}

[project.optional-dependencies]
dev = ["pytest~=7.0", "pytest-mock~=3.0", "responses~=0.24.0", "black>=23.11.0", "isort>=5.12.0"]

[project.scripts]
launch = "launch.cli.entrypoint:cli"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = [
    "test/unit"
]
